Importing mods after testing the code on the desktop
I messed up slightly github.  The version on github lagged behind
the one on the laptop where I did most of the development.  I now
patched up the github version but from the desktop (don't ask how).

The code works on the desktop.  Currently, there is a mismatch in the
gnuplot binary location between the deskto pand laptop, which needs
to be fixed (on the laptop).  The code also needs to be tested on SBCL.