Changed command echoing, and other cosmetic fixes
The previous echoing setup would accumulate input until I
would read-out the stored gnuplot commands.

This resulted in accumulating large amounts of data in the
echo stream.  I discovered this when I was plotting large
data-sets.  After a few plots I received an error in clisp
that the maximum string length (hundreds of millions of characters)
was exceeded.

I replaced this system with a macro that creates an echo stream on
demand, and automatically reads the data into a special variable,
emptying the stream.

Elimination of the automitic echo stream resulted in considerable
increase in plotting speed